W-073 - move dto to other package
