+++ backend/plortal/src/main/java/com/plortal/plortal/model/dto/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.plortal.plortal.dto;
+package com.plortal.plortal.model.dto;
 
 public record LoginRequest(
         String email,
