+++ backend/plortal/src/main/java/com/plortal/plortal/service/impl/NewsServiceImpl.java
@@ -1,8 +1,8 @@
 package com.plortal.plortal.service.impl;
 
 import com.plortal.plortal.model.entity.News;
-import com.plortal.plortal.dto.NewsDTO;
-import com.plortal.plortal.mapper.NewsDTOMapper;
+import com.plortal.plortal.model.dto.NewsDTO;
+import com.plortal.plortal.model.mapper.NewsDTOMapper;
 import com.plortal.plortal.repository.NewsRepository;
 import com.plortal.plortal.service.NewsService;
 import org.springframework.beans.factory.annotation.Autowired;
